#4 adicionado listagem dos livros na home.zul e função de reservar livro

+++ src/main/java/br/com/jkuhn/library/dao/IBookDAO.java
@@ -3,6 +3,6 @@ package br.com.jkuhn.library.dao;
 import br.com.jkuhn.library.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IBookDAO extends JpaRepository<Book, Integer> {
+public interface IBookDAO extends JpaRepository<Book, Long> {
     public Book findByName(String name);
 }
